refactor: 엔티티 웹툰 구조 변경
