+++ src/entities/search/model/store.ts
@@ -1,6 +1,6 @@
 import { create } from "zustand";
 import { SearchState } from "../ui/types.ts";
-import {WebtoonPaginatedResponse} from "@/entities/webtoon/ui/types.ts";
+import {WebtoonPaginatedResponse} from "@/entities/webtoon/model/types.ts";
 
 const emptyResults: WebtoonPaginatedResponse = {
   content: []
