+++ src/pages/main/Main.tsx
@@ -5,7 +5,7 @@ import CharacterChat from "@/features/character-chat/ui/CharacterChat.tsx";
 import CommunityReviews from "@/features/webtoon-reviews/ui/CommunityReviews.tsx";
 import AIAnalysisBanner from "@/features/ai-banner/ui/AIAnalysisBanner.tsx";
 import { useUserStore } from '@/entities/auth/model/userStore.ts';
-import { useWebtoonStore } from '@/entities/webtoon/model/store.ts';
+import { useWebtoonStore } from '@/entities/webtoon/api/store.ts';
 
 const Main: React.FC = () => {
     const [isAuthenticated, setIsAuthenticated] = useState(useUserStore.getState().isAuthenticated);
