+++ src/pages/search/WebtoonSearchResults.tsx
@@ -1,4 +1,4 @@
-import { WebtoonInfo, WebtoonPaginatedResponse } from "@/entities/webtoon/ui/types.ts"
+import { WebtoonInfo, WebtoonPaginatedResponse } from "@/entities/webtoon/model/types.ts"
 import { useSearchStore } from "@/entities/search/model/store.ts"
 import { useEffect, useState } from "react"
 import { useLocation } from "react-router-dom"
